+++ frontend/stylesheets/_spotlight.less
@@ -179,10 +179,6 @@
   }
 }
 
-#add-missing-dataset-image {
-  background-image: url("/assets/images/add-missing-dataset-thumbnail.jpg");
-}
-
 .mini-dataset-thumbnail-grid {
   display: grid;
   grid-gap: 10px;
@@ -207,44 +203,6 @@
   }
 }
 
-.spotlight-hero-button {
-  margin-right: 50px;
-  margin-bottom: 30px;
-  display: inline-block;
-  color: rgba(255, 255, 255, 0.8);
-
-  &:hover {
-    color: rgba(255, 255, 255, 1);
-  }
-}
-
-#oxalis-header {
-  margin-bottom: 20px;
-  margin-top: 20px;
-  text-align: center;
-  height: 100px;
-  vertical-align: middle;
-  display: inline-block;
-  font-size: 80px;
-  line-height: normal;
-  color: #000;
-  margin-left: 30px;
-
-  img {
-    height: 100px;
-    width: 100px;
-  }
-
-  @media @smartphones {
-    font-size: 2rem;
-
-    img {
-      height: 60px;
-      width: 60px;
-    }
-  }
-}
-
 .container-with-hidden-icon {
   .hidden-icon {
     visibility: hidden;
@@ -254,11 +212,6 @@
       visibility: visible;
     }
   }
-}
-
-#spotlight-footnote {
-  text-align: center;
-  margin: 30px 0;
 }
 
 .spotlight-registration-form {
