+++ frontend/stylesheets/antd_overwrites.less
@@ -115,14 +115,10 @@ label.ant-checkbox-wrapper {
 
 // Remove Chinese Quote font, which is messing up our quotes
 .ant-alert {
-  font-family: "Titillium Web", "Monospaced Number", -apple-system, BlinkMacSystemFont, "Segoe UI",
-    Roboto, "PingFang SC", "Hiragino Sans GB", "Microsoft YaHei", "Helvetica Neue", Helvetica, Arial,
-    sans-serif;
+  font-family: @font-family;
 }
 .ant-card {
-  font-family: "Titillium Web", "Monospaced Number", -apple-system, BlinkMacSystemFont, "Segoe UI",
-    Roboto, "PingFang SC", "Hiragino Sans GB", "Microsoft YaHei", "Helvetica Neue", Helvetica, Arial,
-    sans-serif;
+  font-family: @font-family;
 }
 
 .antd-no-border-list .ant-list-item {
