+++ frontend/stylesheets/trace_view/_tracing_view.less
@@ -227,20 +227,42 @@
   }
 }
 
-// Dark mode for UI within navbar and footer
-@dark-bg: #383d48;
-@dark-fg: #f1f1f1;
-@dark-border: #585868;
-@primary-bg: #2174c1;
-@primary-border: #1890ff;
-@hover-border: #40a9ff;
+.navbar-header,
+.statusbar-footer {
+  // Dark mode for UI within navbar and footer
+  @dark-bg: #383d48;
+  @dark-fg: #f1f1f1;
+  @dark-border: #585868;
+  @primary-bg: #2174c1;
+  @primary-border: #1890ff;
+  @hover-border: #40a9ff;
 
-.ant-layout-header,
-.ant-layout-footer {
-  width: 100%;
+  background-color: @layout-header-background;
+  color: rgba(255, 255, 255, 0.67);
   z-index: 1000;
-  color: rgba(255, 255, 255, 0.67) !important;
-  background: #001529 !important;
+
+  &.statusbar-footer {
+    left: 0px;
+    bottom: 0px;
+    right: 0px;
+    position: absolute;
+    height: @footer-height;
+    overflow: hidden;
+    font-size: 11px;
+    text-align: center;
+    padding: 0px;
+  }
+
+  &.navbar-header {
+    width: 100%;
+  }
+
+  .navbar-separator {
+    height: 100%;
+    margin-left: 2px;
+    margin-right: 10px;
+    border-left: 1px #666879 solid;
+  }
 
   .ant-select-arrow,
   .ant-badge-dot {
@@ -345,33 +367,21 @@
       border-color: @primary-border !important;
     }
   }
-}
 
-.footer {
-  left: 0px;
-  bottom: 0px;
-  right: 0px;
-  position: absolute;
-  height: @footer-height;
-  overflow: hidden;
-  font-size: 11px;
-  text-align: center;
-  padding: 0px;
-}
-
-.center-item-using-flex {
-  display: flex !important;
-  align-items: center;
-  justify-content: center;
-}
-
-.no-hover-highlighting {
-  &:hover,
-  &:focus,
-  &:active {
-    border-color: initial;
-    background-color: initial;
-    color: initial;
+  .left-border-button.footer-button,
+  .right-border-button.footer-button {
+    position: absolute;
+    top: 0px;
+    bottom: 0px;
+    height: auto;
+    color: @dark-fg;
+    background-color: inherit;
+  }
+  .left-border-button.footer-button {
+    left: 8px;
+  }
+  .right-border-button.footer-button {
+    right: 8px;
   }
 }
 
@@ -386,22 +396,6 @@
   }
 }
 
-.left-border-button.footer-button,
-.right-border-button.footer-button {
-  position: absolute;
-  top: 0px;
-  bottom: 0px;
-  height: auto;
-  color: @dark-fg;
-  background-color: inherit;
-}
-.left-border-button.footer-button {
-  left: 8px;
-}
-.right-border-button.footer-button {
-  right: 8px;
-}
-
 .flex-layout-container {
   left: 0px;
   top: 0px;
@@ -409,4 +403,20 @@
   height: calc(100% - @footer-height);
   position: absolute;
   overflow: hidden;
+}
+
+.no-hover-highlighting {
+  &:hover,
+  &:focus,
+  &:active {
+    border-color: initial;
+    background-color: initial;
+    color: initial;
+  }
+}
+
+.center-item-using-flex {
+  display: flex !important;
+  align-items: center;
+  justify-content: center;
 }
